☕ Refactored navigator from router base folder into separate navigator
