+++ lib/router/navigator/auth.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
 // 🏘️ Local imports
-import 'router.dart';
+import '../router.dart';
 
 class AuthNavigator extends StatelessWidget {
   const AuthNavigator({Key? key}) : super(key: key);
