+++ lib/router/routes.dart
@@ -1,8 +1,8 @@
 
 // 🏘️ Local imports
 import '../screens/auth/resetPassword.dart';
-import 'auth.dart';
-import 'roles.dart';
+import 'navigator/auth.dart';
+import 'navigator/roles.dart';
 import '../../screens/auth/login.dart';
 import '../../screens/auth/onboarding.dart';
 import '../../screens/auth/register.dart';
