+++ lib/screens/roles/admin/home.dart
@@ -7,7 +7,7 @@ import 'package:provider/provider.dart';
 import '../../../core/models/user.dart';
 import '../../../styles/icons/chap_chap_icons.dart';
 import '../../../core/providers/appdata.dart';
-import '../../../router/roles.dart';
+import '../../../router/navigator/roles.dart';
 import '../../../styles/ui/colors.dart';
 import 'items.dart';
