+++ lib/screens/shared/settings.dart
@@ -7,7 +7,7 @@ import 'package:provider/provider.dart';
 
 // 🏘️ Local imports
 import '../../core/providers/user.dart';
-import '../../router/roles.dart';
+import '../../router/navigator/roles.dart';
 import '../../styles/icons/chap_chap_icons.dart';
 import '../../core/models/garage.dart';
 import '../../core/providers/appdata.dart';
